🚚 Rename: 사이드바 features로 이동
